requirements.txt
langchain==0.1.13
openai==1.14.2
streamlit==1.24.0
langchain-openai==0.1.0
python-dotenv==1.0.1
